+++ src/Core/SFX/audioCore.java
@@ -10,6 +10,9 @@ import java.net.URL;
  * KM
  * May 18 2017
  * Handles the loading and playing of audio files.
+ *
+ * SOURCES:
+ *
  */
 
 public class audioCore extends Thread {
@@ -18,18 +21,36 @@ public class audioCore extends Thread {
     double volume;
     double delay;
     double duration;
+    boolean loop = false;
 
-    static { JFXPanel fxPanel = new JFXPanel(); }
+    private Media media;
+    private MediaPlayer mediaPlayer;
+
+    static { JFXPanel fxPanel = new JFXPanel(); } //needed to play audio, apparently
 
     @Override
     public void run() {
 
-        URL file = this.getClass().getResource("Resources/" + this.sound);
+        initializeAudioData(); //sets up the audio file
 
+        if (loop) {
+            loopAudio(); //if loop is enabled, loop the audio
+        } else {
+            playAudio(); //otherwise, just play the audio file
+        }
+
+        System.out.println("Audio thread closed.");
+
+        Thread.currentThread().interrupt(); //closes the thread down
+    }
+
+    private void initializeAudioData() {
+
+        URL file = this.getClass().getResource("Resources/" + this.sound);
         System.out.println("Playing audio file : " + file);
 
-        final Media media = new Media(file.toString());
-        final MediaPlayer mediaPlayer = new MediaPlayer(media);
+        media = new Media(file.toString());
+        mediaPlayer = new MediaPlayer(media);
 
         try { //pause and wait for the mediaplayer to initialize
             Thread.sleep((long)delay);                 //1000 milliseconds is one second.
@@ -42,21 +63,32 @@ public class audioCore extends Thread {
         if (this.duration == 0) {
             duration = mediaPlayer.getTotalDuration().toMillis();
         }
-        long longDuration = (new Double(duration)).longValue();
+
+    }
+
+    private void playAudio() {
 
         mediaPlayer.play(); //plays the audio
 
         try {
-            System.out.println("Closing audio thread when music ends in " + (longDuration / 1000) + " s");
-            Thread.sleep(longDuration);                 //1000 milliseconds is one second.
+            System.out.println("Closing audio thread when music ends in " + ((long)duration / 1000) + " s");
+            Thread.sleep((long)duration);                 //1000 milliseconds is one second.
         } catch(InterruptedException ex) {
             Thread.currentThread().interrupt();
         }
         mediaPlayer.stop();
-        System.out.println("Audio thread closed.");
 
-        Thread.currentThread().interrupt(); //closes the thread down
     }
+
+    private void loopAudio() {
+
+        mediaPlayer.setCycleCount(MediaPlayer.INDEFINITE);
+        mediaPlayer.play();
+
+    }
+
+
+
 
     public audioCore(String toPlay, double volumeLevel) {
         this.sound = toPlay;
@@ -79,6 +111,15 @@ public class audioCore extends Thread {
         this.volume = volumeLevel;
         this.delay = delay;
         this.duration = duration;
+
+    }
+
+    public audioCore(String toPlay, double volumeLevel, boolean loop) {
+        this.sound = toPlay;
+        this.volume = volumeLevel;
+        this.delay = 1000;
+        this.duration = 0;
+        this.loop = loop;
 
     }